python/pyproject.toml
[project]
name = "ai-agent"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "agent-framework==1.0.0b251120",
    "asyncio>=4.0.0,<5.0",
    "azure-identity>=1.25.1,<2.0",
    "lagom>=2.7.7,<3.0",
    "pydantic-settings>=2.12.0,<3.0",
    "pydantic>=2.12.4,<3.0",
    "python-dotenv>=1.2.1,<2.0",
]

[dependency-groups]
cli = [
    "pyright>=1.1.407,<2.0", 
    "ruff>=0.14.7,<0.15.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["src/my_agent"]
